+++ CHANGES.rst
@@ -9,6 +9,9 @@
 - TEST CLEANUP
     - genometracks
 
+- TEST CLEANUP
+    - plotting/clustering
+
 Changes to notebooks
 ^^^^^^^^^^^^^^^^^^^^
 - prepare_for_cellxgene: Adapt to new mampok verison 2.0.9